gpu/vulkan: fix C4267 size_t->uint32_t narrowing in PM4 trace (restore host /WX build)
vulkan_command_processor.cc had two std::min<uint32_t>(countof(...), length/N)
calls whose size_t arguments narrow to uint32_t, tripping C4267 - treated as an
error under the host Windows /WX build (xenia-gpu-vulkan), pre-existing since
8cc760e0cf (2026-05-18). This broke the host xenia-app desktop build, blocking
the host compile-check path for GPU/app changes (gradle/arm64/clang was
unaffected - it doesn't /WX these), as hit this session while verifying the
save-state work.

Switch the inner min to std::min<size_t> (both args are already size_t -> no
narrowing) and keep the explicit outer uint32_t() cast (the result is <= the
small countof, always fits). Behavior identical. Restores the host /WX build's
xenia-gpu-vulkan compile.

Build-verified host x64 (xenia-gpu-vulkan) + arm64 gradle.

+++ src/xenia/gpu/vulkan/vulkan_command_processor.cc
@@ -1743,7 +1743,7 @@ bool VulkanCommandProcessor::ReadbackSharedMemoryRange(uint32_t address,
         }
         uint32_t first_words[8] = {};
         uint32_t first_word_count =
-            uint32_t(std::min<uint32_t>(xe::countof(first_words), length / 4));
+            uint32_t(std::min<size_t>(xe::countof(first_words), length / 4));
         for (uint32_t i = 0; i < first_word_count; ++i) {
           std::memcpy(&first_words[i], bytes + i * sizeof(uint32_t),
                       sizeof(uint32_t));
@@ -1988,7 +1988,7 @@ void VulkanCommandProcessor::TraceVertexFetchSources(
           ++nonzero_samples;
         }
       }
-      uint32_t first_word_count = uint32_t(std::min<uint32_t>(
+      uint32_t first_word_count = uint32_t(std::min<size_t>(
           xe::countof(first_words_raw), length / sizeof(uint32_t)));
       for (uint32_t i = 0; i < first_word_count; ++i) {
         std::memcpy(&first_words_raw[i], bytes + i * sizeof(uint32_t),
